Create Small World: Disney, Marvel, DC, and Reality
My Game: Small World; Disney, Marvel and DC

I would love to create more interactive story Building and Collecting games, Like Pokémon and sky landers, but using Disney, Marvel and DC comics characters and background/ environments. In combination with  general real world history, and events that gave/give shape to every thing( 9-11-2001, WW2) and the building blocks of education in a world collecting and building game that uses both physical and digital platforms…In other words to create a game that is all around educational collectable buildable and timeless and age none restrictive….

a) What option did you pick? I chose a combination of challenges 3: when I find you, and 4: pick it up, and 2; it is mine.

b) What is the theme and goal of your idea? Can my game have more than one theme? In that it is collectable as you acquire land, characters, stories and knowledge( history, math, geography and cultural understanding) and levels as one plays.

c) What format is this game? (e.g., board game, video game, etc.) The easiest way to play is digital, physically it is akin to trading card games but with books and action figures( that are the characters in the books) instead of cards and Monopoly/ Life where in the money is fact cards, figurines(Babies), books, and fact sheets and building sets(Legos).

e) Identify the avatars/players. (How are players represented? Are they?) when it comes to the physical the player is their own avatar, in the digital version they start out as either the main character of their first story/book or a digitized self-portrait with wardrobe changes with the seasons region locked (and if possible time zone oriented clock) and achievements. The  physical game board is a map of the location indicated in the story(s), such as France for Beauty and the Beast, the Hunchback of Notre Dame and Ratatouille and the buildings are Lego play houses/castles and vehicles( with size appropriate figures) for board game play. There will be a solo non chess like play in that the human characters are 12 inch clothing changeable (per the story/script which is attached in trading card form)babies  and modern/film release dated close sold with(and fact card included) and the animals such as the Dalmatians or the Oristocats come as 24 inch and under(depending on indicated age, Lion King) possible dolls

f) Identify conflict between players in the game play. (How do players interact with each other?) To trade and help build/collect/ advance or to go it alone that is the question. In the digital version players are to share the wealth and communicate in order to make friends and time differential ( the leading conflict of Farmville, Disney Kingdoms, and Ever Merge and the like).. 

g) Who is the audience for this game? any one over 3 who likes animals, Universal studios films(strictly PG14 and calmer ), or the classics of literature,  Disney, Marvel and DC, and comics, and facts/learning  and real world/global history, and events that gave/give shape to every thing, and seeing how they combine into one great big narrative of life.

elfquest-game6 rules.pdf
(10.81 MB)
elfquest-game5.pdf
(4.54 MB)
elfquest-game4.jpg.crdownload.webp
(70.01 KB)
elfquest-game3.jpg.crdownload
(284.21 KB)
ElfQuest game2.jpg
(79.17 KB)
ElfQuest game1.jpg
(85.18 KB)
disney-gargoyles-awakening-board-game-1.jpeg
(160.78 KB)
disney-gargoyles-awakening-board-game-8.jpeg
(163.2 KB)
disney-gargoyles-awakening-board-game-7.jpeg
(183.63 KB)
LEGO-Batman-The-Videogame.webp
(81.72 KB)